feat(core): add ObservableCardanoNode interface
hoist some types from CardanNode.ts
