+++ packages/core/package.json
@@ -62,6 +62,14 @@
     "ts-jest": "^28.0.7",
     "typescript": "^4.7.4"
   },
+  "peerDependencies": {
+    "rxjs": "^7.4.0"
+  },
+  "peerDependenciesMeta": {
+    "rxjs": {
+      "optional": true
+    }
+  },
   "dependencies": {
     "@cardano-ogmios/client": "5.6.0",
     "@cardano-ogmios/schema": "5.6.0",
