+++ packages/core/src/CardanoNode/types/index.ts
@@ -1,2 +1,3 @@
 export * from './CardanoNode';
 export * as CardanoNodeErrors from './CardanoNodeErrors';
+export * from './ObservableCardanoNode';
